Button Layout, Reset Position, Personal Machine Config
Added a menu option to reset the position reading to 0 degrees
Changed how the buttons work. From right to left:
Power is now always able to be pressed. So you can stop the leadscrew while the spindle is running
Settings: remains the same
Fwd/rev: remains the same
Feed/thead: combined with IN/mm into one mode button. Press to cycle between feed in mm, thread in mm, feed in inches, thread in inches
In/MM: changed to a rapid left button,  not yet implemented
-: changed to a rapid right button. Not yet implemented
Extra button behind the control panel: changed to -
+: remains the same

Right now I am debating on how/if to implement the rapid function. Current thoughts are that it requires a long press, slowly accelerates to a max speed, and stops when released